Day 17
Had to try a few different things to spot a repeat, comparing the height
diffs worked to make sure we had a consistent repeat. Without that,
there were situations where a long piece would slip past and mess up the
repeat.

Then I had the height offset calculation wrong (but only on the input,
not on the example) because I wasn't using the truncated repeats value.